+++ Service.API/Domain/Domain/Services/FantasyDataService/IFantasyDataService.cs
@@ -10,6 +10,6 @@ namespace Fantasy.API.Domain.Services.FantasyService
     {
         Task<ServiceResult<List<ContestBO>>> GetContestsAsync();
         Task<ServiceResult<List<PlayerBO>>> GetPlayersFromTeamAsync(int teamId);
-        Task<ServiceResult<TeamBO>> GetTeamAsync(int teamId);
+        Task<ServiceResult<TeamBO>> GetTeamAsync(long teamId);
     }
 }
